fix init mu bug
originally we fit mu for each allelic cluster, hence why we picked out the mu
value for the segment we're running on here. Now we fit a mu for each segment,
so we need to switch to the index of our segment to get the correct mu. This only
effected the init_lnp function for the cov mcmc since after that function is run the
lnp mu is used rather than the poisson mu